Adding the writing new file function
So I put in a function to write a new file containing all of the lines from each month. I think it's only going to work for the first 12 months, but it looks as right as I can make it out to be. I put it in this spot so that I could grab the appended lists and copy them before they get reset.
Please test run this as it may crash the whole thing without letting it finish.